requirements.txt
python-dotenv
prefect
prefect-dbt
pandas==1.5.2
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.4
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
sqlalchemy==1.4.46
